+++ lib/angular-tree-component.ts
@@ -17,32 +17,13 @@ import { TreeNodeDropSlot } from './components/tree-node-drop-slot.component';
 import { TreeNodeExpanderComponent } from './components/tree-node-expander.component';
 import { TreeNodeChildrenComponent } from './components/tree-node-children.component';
 import { TreeNodeCollectionComponent } from './components/tree-node-collection.component';
+import { TreeNodeWrapperComponent } from './components/tree-node-wrapper.component';
 import { TreeViewportComponent } from './components/tree-viewport.component';
 import { TreeDropDirective } from './directives/tree-drop.directive';
 import { TreeDragDirective } from './directives/tree-drag.directive';
+import { TreeAnimateOpenDirective } from './directives/tree-animate-open.directive';
 
 import './polyfills';
-import { useStrict } from 'mobx';
-
-/*
-  Turn on to check if tree supports strict mode in MobX.
-  But remember to turn off (to allow users of MobX not to use strict mode in their apps)
-*/
-// useStrict(true);
-
-const exportedDirectives = [
-  TreeComponent,
-  TreeNodeComponent,
-  TreeNodeContent,
-  LoadingComponent,
-  TreeDropDirective,
-  TreeDragDirective,
-  TreeNodeExpanderComponent,
-  TreeNodeChildrenComponent,
-  TreeNodeDropSlot,
-  TreeNodeCollectionComponent,
-  TreeViewportComponent
-];
 
 export {
   TreeModel,
@@ -71,10 +52,34 @@ export {
 
 @NgModule({
   declarations: [
-    ...exportedDirectives
+    TreeComponent,
+    TreeNodeComponent,
+    TreeNodeContent,
+    LoadingComponent,
+    TreeDropDirective,
+    TreeDragDirective,
+    TreeNodeExpanderComponent,
+    TreeNodeChildrenComponent,
+    TreeNodeDropSlot,
+    TreeNodeCollectionComponent,
+    TreeViewportComponent,
+    TreeNodeWrapperComponent,
+    TreeAnimateOpenDirective
   ],
   exports: [
-    ...exportedDirectives
+    TreeComponent,
+    TreeNodeComponent,
+    TreeNodeContent,
+    LoadingComponent,
+    TreeDropDirective,
+    TreeDragDirective,
+    TreeNodeExpanderComponent,
+    TreeNodeChildrenComponent,
+    TreeNodeDropSlot,
+    TreeNodeCollectionComponent,
+    TreeViewportComponent,
+    TreeNodeWrapperComponent,
+    TreeAnimateOpenDirective
   ],
   imports: [
     CommonModule,