+++ lib/models/tree-node.model.ts
@@ -1,11 +1,11 @@
-import { observable, computed, autorun } from 'mobx';
+import { observable, computed, autorun, action } from 'mobx';
 import { TreeModel } from './tree.model';
 import { TreeOptions } from './tree-options.model';
 import { ITreeNode } from '../defs/api';
 import { TREE_EVENTS } from '../constants/events';
 
-const first  = require('lodash/first');
-const last  = require('lodash/last');
+import * as _ from 'lodash';
+const { first, last } = _;
 
 export class TreeNode implements ITreeNode {
   @computed get isHidden() { return this.treeModel.isHidden(this); };
@@ -328,7 +328,7 @@ export class TreeNode implements ITreeNode {
     return this.options.nodeHeight(this);
   }
 
-  _initChildren() {
+  @action _initChildren() {
     this.children = this.getField('children')
       .map((c, index) => new TreeNode(c, this, this.treeModel, index));
   }