+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+<a name="3.4.1"></a>
+# 3.4.1 (2017-29-04)
+* Fixed SystemJS errors
+
+<a name="3.4.0"></a>
+# 3.4.0 (2017-28-04)
+* Add animations
+* Add node wrapper template
+* Fixed drop slot style
+
 <a name="3.3.1"></a>
 # 3.3.1 (2017-19-04)
 * Support useStrict mode in MobX